Use parser from raml-toolkit
