+++ packages/generator/src/templateHelpers.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { model } from "amf-client-js";
+import { model } from "@commerce-apps/raml-toolkit";
 
 import { commonParameterPositions } from "@commerce-apps/core";
